Storage for keys and values are removed
This makes the code for Map.Get more simple because it only returns
the index of the key and not its corresponding value. However, since
the added keys are not stored, Map.Get loses its ability to determine
if a given key is added or not. In fact, Map.Get will always return a
value between 0 and Map.Cap()-1. It is up to the user to validate this
returned value.

One advantage of not having to store values is that the user has
the freedom to specify the type of the map's backing store (ie. not
restricted to byte slices). Because of this, the user doesn't need to
encode values when adding to the map.