Fixed delete bug
If after sorting the items, the first item has its delete flag set
to true, it will not be moved to the front of the array. Hence, it
will not be deleted. To fix this, the pkey variable must be set to
a value not equal to the first item's key before iterating through
the array.